Extract slurm api error handler
