+++ trailblazer/services/analysis_service/error_handler.py
@@ -2,12 +2,7 @@ import logging
 from functools import wraps
 
 from trailblazer.exc import AnalysisServiceError
-from trailblazer.exceptions import (
-    JobServiceError,
-    SlurmAPIServiceError,
-    SlurmCLIServiceError,
-    TowerServiceError,
-)
+from trailblazer.exceptions import JobServiceError
 
 LOG = logging.getLogger(__name__)
 
@@ -18,7 +13,7 @@ def handle_analysis_service_errors(func):
     def wrapper(*args, **kwargs):
         try:
             return func(*args, **kwargs)
-        except (JobServiceError) as error:
+        except JobServiceError as error:
             raise AnalysisServiceError(error) from error
 
     return wrapper
